+++ Assets/_Game/Core/UserState/_Handler/Currencies/ICurrenciesHandler.cs
@@ -6,6 +6,6 @@ namespace _Game.Core.UserState._Handler.Currencies
     public interface ICurrenciesHandler
     {
         void AddCoins(in float quantity, CurrenciesSource source);
-        void AddGems(in int quantity, CurrenciesSource source);
+        void AddGems(in float quantity, CurrenciesSource source);
     }
 }
