+++ Assets/_Game/Core/UserState/_Handler/_Purchase/IPurchaseStateHandler.cs
@@ -6,7 +6,7 @@ namespace _Game.Core.UserState._Handler._Purchase
     public interface IPurchaseStateHandler
     {
         void PurchaseUnit(UnitType type, float price, CurrenciesSource source);
-        void PurchaseCoinsWithGems(int quantity, int price, CurrenciesSource source);
+        void PurchaseCoinsWithGems(float quantity, int price, CurrenciesSource source);
         void AddPurchase(string id);
     }
 }
